refactor(winscard): move container name, certificate, and private key initialization code to the `env` module;
